requirements.txt
# Virtual Autonomous Nanomaterials Lab (VANL) — Dependencies
# Install via: pip install -r requirements.txt

# Numerical Computing
numpy>=1.24.0

# Machine Learning & Optimization
scikit-learn>=1.3.0
scipy>=1.11.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Testing
pytest>=7.0.0

# System Monitoring (for GPU manager)
psutil>=5.9.0

# Security & Encryption
cryptography>=41.0.0

# Database (Phase 5: Enterprise Features)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Caching & Sessions
redis>=5.0.0

# Data Analysis (Phase 4)
pandas>=2.0.0

# Email Validation
email-validator>=2.0.0

# Scheduling & Webhooks (Phase 5 Week 19)
croniter>=2.0.0
aiohttp>=3.9.0

# Report Generation (Phase 5 Week 20)
reportlab>=4.0.0
openpyxl>=3.1.0
python-docx>=1.0.0

# HTTP Requests (for NVIDIA API)
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# NVIDIA ALCHEMI (Optional - for quantum-accurate predictions)
# Install separately: pip install nvalchemi-toolkit ase
# Requires Python 3.11, 3.12, or 3.13 (NOT 3.14+)
torch>=2.0.0
